Just checking in--Separated out much of the KNN-specific information to a separate class, assigning pure virtuals where necessary. Cleaned up a lot of the code.
It's still much slower than a single tree, but the aggressive pruning seems to perform quite well (somehow). The next step the remove the unnecessary calls to compute the centroid and bounding boxes to a separate
tree construction step, then focus on singular component optimization. After rethinking the way the DT traversal works, I'm doubting incremental distance calculations will provide any additional benefit, as the traversal is fundamentally
different from a regular traversal--the query point is not know ahead of time. It may be possible though, still working on it.